Add libffi dependency as Cocoapod.
It took a way longer than I expected.
Basically a lot of files that use to sit at Tests/ now sit at the
root of the project (.ruby-version, Gemfile, Podfile...), also the
Pods directory has moved.

The Podfile contains two more targets for the static library which
point to the podspec (to have access to libffi). There is some
minor changes in the ARCHS, HEADERS and LDFLAGS of the main project
and the testing projects. I hope I haven't broken anything.

There is a hack in the post_install hook of the Podfile to avoid
Cocoapods from overwritting some libffi iOS headers with their
OS X versions. I might need to send this patch to either Cocoapods
or libffi, because I think projects depending on Typhoon might
have problems.

Also: it seems that OCMockito and OCHamcrest have been updated.

(jasperblues/Typhoon#102)